engine: address iteration 11 review feedback
Fix override temp-ID remapping: remap disc_main along with disc_exprs
and set new_base after advancing past remapped pre-expressions, so
subsequent elements' TempArrayElement reads align with hoisted
AssignTemp IDs.

Strengthen per-element hoisting detection for multi-dimensional arrays
by also comparing element 0 vs the last element (not just element 1).
For 2D+ arrays, elements 0 and 1 share the same outer subscript, so
comparing only those two misses scalar args that depend on the outer
dimension.

Widen VectorSortOrder index storage from u16 to usize to avoid
truncation for arrays with more than 65535 elements.

Remove dead total_elements binding from per-element hoisting.

+++ src/simlin-engine/src/vm.rs
@@ -2066,7 +2066,8 @@ impl Vm {
                         let n_dims = input_view.dims.len();
 
                         // Collect (value, 1-based-index) pairs
-                        let mut indexed: SmallVec<[(f64, u16); 32]> = SmallVec::with_capacity(size);
+                        let mut indexed: SmallVec<[(f64, usize); 32]> =
+                            SmallVec::with_capacity(size);
                         let mut indices: SmallVec<[u16; 4]> = smallvec::smallvec![0; n_dims];
                         for i in 0..size {
                             let flat_off = input_view.flat_offset(&indices);
@@ -2077,7 +2078,7 @@ impl Vm {
                                 temp_storage,
                                 context,
                             );
-                            indexed.push((val, (i + 1) as u16));
+                            indexed.push((val, i + 1));
                             increment_indices(&mut indices, &input_view.dims);
                         }
